organized lib code
